Apply suggestions from code review continued
