+++ src/libraries/System.Private.CoreLib/src/System/IO/FileInfo.cs
@@ -85,17 +85,6 @@ namespace System.IO
         public StreamWriter AppendText()
             => new StreamWriter(NormalizedPath, append: true);
 
-#if NET6_0_OR_GREATER
-        public StreamReader OpenText(FileStreamOptions options)
-            => File.OpenText(NormalizedPath, options);
-
-        public StreamWriter CreateText(FileStreamOptions options)
-            => File.CreateText(NormalizedPath, options);
-
-        public StreamWriter AppendText(FileStreamOptions options)
-            => File.AppendText(NormalizedPath, options);
-#endif
-
         public FileInfo CopyTo(string destFileName) => CopyTo(destFileName, overwrite: false);
 
         public FileInfo CopyTo(string destFileName, bool overwrite)
@@ -117,15 +106,6 @@ namespace System.IO
             return fileStream;
         }
 
-#if NET6_0_OR_GREATER
-        public FileStream Create(FileStreamOptions options)
-        {
-            FileStream fileStream = File.Create(NormalizedPath, options);
-            Invalidate();
-            return fileStream;
-        }
-#endif
-
         public override void Delete()
         {
             FileSystem.DeleteFile(FullPath);
@@ -146,17 +126,6 @@ namespace System.IO
 
         public FileStream OpenWrite()
             => new FileStream(NormalizedPath, FileMode.OpenOrCreate, FileAccess.Write, FileShare.None);
-
-#if NET6_0_OR_GREATER
-        public FileStream Open(FileMode mode, FileStreamOptions options)
-            => File.Open(NormalizedPath, mode, options);
-
-        public FileStream OpenRead(FileStreamOptions options)
-            => File.OpenRead(NormalizedPath, options);
-
-        public FileStream OpenWrite(FileStreamOptions options)
-            => File.OpenRead(NormalizedPath, options);
-#endif
 
         // Moves a given file to a new location and potentially a new file name.
         // This method does work across volumes.
